web/requirements.web.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiohttp
python-dotenv
python-multipart
python-jose[cryptography]
pydantic-settings
jinja2